requirements.txt
Flask
Flask-Cors
llama-index-core>=0.11.0
llama-index>=0.11.0
# see https://llamahub.ai for all integrations
llama-index-embeddings-openai  # can replace with the embed model you want to use
llama-index-llms-openai  # can replace with the llm you want to use
llama-index-agent-openai
llama-index-readers-web
llama-parse
panda
matplotlib
gradio
llama-index-utils-workflow
llama-index-readers-whisper
llama-index-utils-workflow
